+++ HopInNow/app/src/main/AndroidManifest.xml
@@ -17,8 +17,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".DriverProfileActivity"></activity>
-        <activity android:name=".RiderProfileActivity" />
+        <activity android:name=".Activities.SignupAndLogin.DriverProfileActivity"></activity>
+        <activity android:name=".Activities.SignupAndLogin.RiderProfileActivity" />
         <activity android:name=".UserProfileActivity" />
 
         <uses-library
@@ -29,7 +29,7 @@
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/map_key" />
 
-        <activity android:name=".SignupActivity">
+        <activity android:name=".Activities.SignupAndLogin.SignupActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <action android:name="android.intent.action.VIEW" />
@@ -37,7 +37,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".RiderMapActivity" />
+        <activity android:name=".Activities.Map.RiderMapActivity" />
     </application>
 
 </manifest>
