+++ HopInNow/app/src/main/java/com/example/hopinnow/Activities/SignupAndLogin/DriverProfileActivity.java
@@ -1,8 +1,10 @@
-package com.example.hopinnow;
+package com.example.hopinnow.Activities.SignupAndLogin;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.example.hopinnow.R;
 
 public class DriverProfileActivity extends AppCompatActivity {
