+++ HopInNow/app/src/main/java/com/example/hopinnow/Activities/SignupAndLogin/SignupActivity.java
@@ -1,10 +1,14 @@
-package com.example.hopinnow;
+package com.example.hopinnow.Activities.SignupAndLogin;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 import android.view.View;
 import android.widget.EditText;
+
+import com.example.hopinnow.Database.DatabaseAccessor;
+import com.example.hopinnow.R;
+import com.example.hopinnow.entities.Rider;
 
 public class SignupActivity extends AppCompatActivity {
     public static final String TAG = "SignupActivity";
@@ -33,6 +37,6 @@ public class SignupActivity extends AppCompatActivity {
         String email = emailEditText.getText().toString();
 
         Rider rider = new Rider(name, password, phoneNumber, email,null,null);
-        databaseAccessor.registerRider(getApplicationContext(), rider);
+        databaseAccessor.registerRider(getApplicationContext(), RiderProfileActivity.class, rider);
     }
 }
