--- HopInNow/app/src/main/java/com/example/hopinnow/EmergencyCall.java
@@ -1,26 +1,0 @@
-package com.example.hopinnow;
-
-public class EmergencyCall {
-    private String dialogNumber;
-
-
-    //Constructor
-    public EmergencyCall() {
-        this.dialogNumber = "911";
-    }
-
-    public EmergencyCall(String dialogNumber) {
-        this.dialogNumber = dialogNumber;
-    }
-
-    //getter
-    public String getDialogNumber() {
-        return dialogNumber;
-    }
-
-    //setter
-    public void setDialogNumber(String dialogNumber) {
-        this.dialogNumber = dialogNumber;
-    }
-}
-
