--- HopInNow/app/src/main/java/com/example/hopinnow/RiderProfileActivity.java
@@ -1,14 +1,0 @@
-package com.example.hopinnow;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class RiderProfileActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_rider_profile);
-    }
-}
