--- HopInNow/app/src/main/java/com/example/hopinnow/Trip.java
@@ -1,57 +1,0 @@
-package com.example.hopinnow;
-
-import android.location.Location;
-
-import com.google.android.gms.maps.model.LatLng;
-
-import java.util.Date;
-
-public class Trip extends Ride {
-    private Double cost;
-    private Double rating;
-    private Date dropOffDateTime;
-    private int duration;
-
-    public Trip(){}
-
-    public Trip(Driver driver, Rider rider, LatLng pickUpLoc, LatLng dropOffLoc, String pickUpLocName, String dropOffLocName, Date pickUpDateTime,
-                Date dropOffTime, int duration, Car car, Double cost, Double rating) {
-        super(driver, rider, pickUpLoc, dropOffLoc, pickUpLocName, dropOffLocName, pickUpDateTime, car);
-        this.cost = cost;
-        this.rating = rating;
-        this.dropOffDateTime = dropOffTime;
-        this.duration = duration;
-    }
-
-    public Double getCost() {
-        return cost;
-    }
-
-    public void setCost(Double cost) {
-        this.cost = cost;
-    }
-
-    public void setDropOffTime(Date dropOffTime) {
-        this.dropOffDateTime = dropOffTime;
-    }
-
-    public void setDuration(int duration) {
-        this.duration = duration;
-    }
-
-    public int getDuration() {
-        return duration;
-    }
-
-    public Date getDropOffTime() {
-        return dropOffDateTime;
-    }
-
-    public Double getRating() {
-        return rating;
-    }
-
-    public void setRating(Double rating) {
-        this.rating = rating;
-    }
-}
