+++ HopInNow/app/src/main/java/com/example/hopinnow/entities/EmergencyCall.java
@@ -1,0 +1,26 @@
+package com.example.hopinnow.entities;
+
+public class EmergencyCall {
+    private String dialogNumber;
+
+
+    //Constructor
+    public EmergencyCall() {
+        this.dialogNumber = "911";
+    }
+
+    public EmergencyCall(String dialogNumber) {
+        this.dialogNumber = dialogNumber;
+    }
+
+    //getter
+    public String getDialogNumber() {
+        return dialogNumber;
+    }
+
+    //setter
+    public void setDialogNumber(String dialogNumber) {
+        this.dialogNumber = dialogNumber;
+    }
+}
+
