+++ HopInNow/app/src/main/java/com/example/hopinnow/entities/EstimateFare.java
@@ -1,0 +1,39 @@
+package com.example.hopinnow.entities;
+
+import com.google.android.gms.maps.model.LatLng;
+
+import java.util.Date;
+
+public class EstimateFare {
+
+    public EstimateFare(){}
+
+    //estimates fair fare based on distance and time
+    //Base Fare + ((Cost per minute x time of the ride)
+    //      + (cost per mile x ride distance) x surge boost multiplier)
+    //      + booking fee = Passengers Ride Fare.
+    /**give attributes locations or google map object?*/
+    public Double estimateFare(LatLng pickUpLoc, LatLng dropOffLoc, Date pickUpDateTime){
+        Double price;
+        Double baseFare = 1.5;
+        Double costPerMinute = 0.16;
+        Double costPerMile = 0.8;
+
+        /**calls google map route calculation
+         *      returns distance, time needed
+         * distance, time = GOOGLE MAP METHOD that returns List<Double>
+         * */
+
+        //temporary, to be replaced by above
+        LatLng pickUp = pickUpLoc;
+        LatLng dropOff = dropOffLoc;
+        Date dateTime = pickUpDateTime;
+        Double distance = 1.0;
+        Double time = 1.0; //= pickUpDateTime;
+
+
+        price = baseFare + (costPerMinute * time) + (distance * costPerMile);
+
+        return price;
+    }
+}
