+++ HopInNow/app/src/main/res/layout/activity_driver_profile.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DriverProfileActivity">
+    tools:context=".Activities.SignupAndLogin.DriverProfileActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
