+++ HopInNow/app/src/main/res/layout/activity_map.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".RiderMapActivity">
+    tools:context=".Activities.Map.RiderMapActivity">
 
     <fragment
         android:id="@+id/map"
