+++ HopInNow/app/src/main/res/layout/activity_signup.xml
@@ -9,7 +9,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:context=".SignupActivity">
+        tools:context=".Activities.SignupAndLogin.SignupActivity">
 
         <Button
             android:id="@+id/button2"
